requirements.txt
asttokens==3.0.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
fonttools==4.56.0
fsspec==2024.6.1
importlib-metadata==8.6.1
importlib-resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
jedi==0.19.2
jinja2==3.1.4
jupyter-client==8.6.3
jupyter-core==5.7.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.4
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.3
opencv-python==4.11.0.86
packaging==24.2
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt-toolkit==3.0.50
psutil==7.0.0
pure-eval==0.2.3
pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pywin32==308
pyzmq==26.2.1
scipy==1.13.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0+cu126
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tornado==6.4.2
traitlets==5.14.3
typing-extensions==4.12.2
wcwidth==0.2.13
zipp==3.21.0